fix Float to Float64

+++ src/utils/costs.jl
@@ -12,7 +12,7 @@ export compute_station_pairwise_costs, read_routing_costs_from_segments
 compute_station_pairwise_costs(candidate_stations::DataFrame) -> Dict::Tuple{Int, Int}, Float64}
 matrix between candidate stations based on Haversine distance.
 """
-function compute_station_pairwise_costs(candidate_stations::DataFrame, walking_speed::Float=1.4)::Dict{Tuple{Int, Int}, Float64}
+function compute_station_pairwise_costs(candidate_stations::DataFrame, walking_speed::Float64=1.4)::Dict{Tuple{Int, Int}, Float64}
     n = nrow(candidate_stations)
     costs = Dict{Tuple{Int, Int}, Float64}()
